Add a person's testimonies to their profile
